+++ scanner/src/main/java/ca/on/oicr/gsi/runscanner/scanner/processor/UltimaApiClient.java
@@ -27,6 +27,11 @@ public class UltimaApiClient {
       new LatencyHistogram(
           "runscanner_ultima_fetch_all_runs_time", "The time to query the NEXUS API for all runs.");
 
+  private static final LatencyHistogram nexus_control_metric_api_fetch_time =
+      new LatencyHistogram(
+          "runscanner_ultima_fetch_run_metric_time",
+          "The time to query the NEXUS API for run metrics.");
+
   private static final LatencyHistogram sample_db_fetch_time =
       new LatencyHistogram(
           "runscanner_ultima_fetch_sample_db",
@@ -118,6 +123,43 @@ public class UltimaApiClient {
               + response.getBody());
     }
     return nodes;
+  }
+
+  public JsonNode fetchBarcodeMetrics(String runId, String barcode) throws IOException {
+
+    HttpEntity<String> entity = new HttpEntity<>(headersNexus);
+    String url =
+        String.format("%s/api/data/metrics/%s/%s?purpose=qtable", apiUrlNexus, runId, barcode);
+
+    ResponseEntity<JsonNode> response;
+    try (AutoCloseable latency_nexus_metric = nexus_control_metric_api_fetch_time.start()) {
+      response = restTemplate.exchange(url, HttpMethod.GET, entity, JsonNode.class);
+      if (!response.getStatusCode().is2xxSuccessful()) {
+        api_nexus_fetch_failed.inc();
+        throw new IllegalStateException(
+            "API request to "
+                + url
+                + " failed, unexpected response status: "
+                + response.getStatusCode());
+      }
+    } catch (Exception e) {
+      api_nexus_fetch_failed.inc();
+      throw new IOException(e);
+    }
+    if (response.getBody() != null) {
+      if (response.getBody().isArray()) {
+        return response.getBody().get(0); // API returns a list of 1
+      } else {
+        return response.getBody();
+      }
+    } else {
+      api_nexus_fetch_failed.inc();
+      throw new IOException(
+          "Unexpected JSON response format from API request to "
+              + url
+              + " : "
+              + response.getBody());
+    }
   }
 
   public JsonNode fetchSampleDB(String samplePlateId) throws IOException {