+++ RELEASE_NOTES.md
@@ -1,11 +1,10 @@
 # Unreleased
 
 Changes:
-* Add "," to Ultima consumables (1000 -> 1,000) and reduce Total Beads to 6 decimal precision
-
-Changes:
 * Add Sample DB support for Ultima
 * Add consumables to Ultima Runs
+* Add "," to Ultima metrics (1000 -> 1,000) and reduce Total Beads to 6 decimal precision
+* Add "Control Sample Bases > Q30 %" Ultima metric
 
 # 2.5.1